Remove update mapping code
