+++ elasticsearch/src/main/java/com/gentics/mesh/search/index/microschema/MicroschemaTransformer.java
@@ -1,8 +1,6 @@
 package com.gentics.mesh.search.index.microschema;
 
-import static com.gentics.mesh.search.index.MappingHelper.DESCRIPTION_KEY;
 import static com.gentics.mesh.search.index.MappingHelper.NAME_KEY;
-import static com.gentics.mesh.search.index.MappingHelper.trigramStringType;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
@@ -32,11 +30,5 @@ public class MicroschemaTransformer extends AbstractTransformer<MicroschemaConta
 		return document;
 	}
 
-	@Override
-	public JsonObject getMappingProperties() {
-		JsonObject props = new JsonObject();
-		props.put(NAME_KEY, trigramStringType());
-		props.put(DESCRIPTION_KEY, trigramStringType());
-		return props;
-	}
+
 }
