+++ elasticsearch/src/main/java/com/gentics/mesh/search/index/project/ProjectTransformer.java
@@ -1,7 +1,6 @@
 package com.gentics.mesh.search.index.project;
 
 import static com.gentics.mesh.search.index.MappingHelper.NAME_KEY;
-import static com.gentics.mesh.search.index.MappingHelper.trigramStringType;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
@@ -30,10 +29,4 @@ public class ProjectTransformer extends AbstractTransformer<Project> {
 		return document;
 	}
 
-	@Override
-	public JsonObject getMappingProperties() {
-		JsonObject props = new JsonObject();
-		props.put(NAME_KEY, trigramStringType());
-		return props;
-	}
 }
