+++ elasticsearch/src/main/java/com/gentics/mesh/search/index/role/RoleIndexHandler.java
@@ -15,6 +15,7 @@ import com.gentics.mesh.core.data.search.SearchQueue;
 import com.gentics.mesh.core.data.search.UpdateDocumentEntry;
 import com.gentics.mesh.graphdb.spi.Database;
 import com.gentics.mesh.search.SearchProvider;
+import com.gentics.mesh.search.index.MappingProvider;
 import com.gentics.mesh.search.index.entry.AbstractIndexHandler;
 
 /**
@@ -25,6 +26,9 @@ public class RoleIndexHandler extends AbstractIndexHandler<Role> {
 
 	@Inject
 	RoleTransformer transformer;
+
+	@Inject
+	RoleMappingProvider mappingProvider;
 
 	@Inject
 	public RoleIndexHandler(SearchProvider searchProvider, Database db, BootstrapInitializer boot, SearchQueue searchQueue) {
@@ -57,6 +61,11 @@ public class RoleIndexHandler extends AbstractIndexHandler<Role> {
 	}
 
 	@Override
+	protected MappingProvider getMappingProvider() {
+		return mappingProvider;
+	}
+
+	@Override
 	public Map<String, String> getIndices() {
 		return Collections.singletonMap(Role.TYPE, Role.TYPE);
 	}
@@ -68,8 +77,7 @@ public class RoleIndexHandler extends AbstractIndexHandler<Role> {
 
 	@Override
 	public RootVertex<Role> getRootVertex() {
-		return boot.meshRoot()
-				.getRoleRoot();
+		return boot.meshRoot().getRoleRoot();
 	}
 
 }
