+++ elasticsearch/src/main/java/com/gentics/mesh/search/index/role/RoleTransformer.java
@@ -1,7 +1,6 @@
  package com.gentics.mesh.search.index.role;
 
 import static com.gentics.mesh.search.index.MappingHelper.NAME_KEY;
-import static com.gentics.mesh.search.index.MappingHelper.trigramStringType;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
@@ -36,14 +35,5 @@ public class RoleTransformer extends AbstractTransformer<Role> {
 		return document;
 	}
 
-	/**
-	 * Return the type specific mapping properties.
-	 */
-	@Override
-	public JsonObject getMappingProperties() {
-		JsonObject props = new JsonObject();
-		props.put(NAME_KEY, trigramStringType());
-		return props;
-	}
 
 }
