+++ elasticsearch/src/main/java/com/gentics/mesh/search/index/schema/SchemaTransformer.java
@@ -2,7 +2,6 @@ package com.gentics.mesh.search.index.schema;
 
 import static com.gentics.mesh.search.index.MappingHelper.DESCRIPTION_KEY;
 import static com.gentics.mesh.search.index.MappingHelper.NAME_KEY;
-import static com.gentics.mesh.search.index.MappingHelper.trigramStringType;
 
 import javax.inject.Inject;
 import javax.inject.Singleton;
@@ -32,12 +31,6 @@ public class SchemaTransformer extends AbstractTransformer<SchemaContainer> {
 		return document;
 	}
 
-	@Override
-	public JsonObject getMappingProperties() {
-		JsonObject props = new JsonObject();
-		props.put(NAME_KEY, trigramStringType());
-		props.put(DESCRIPTION_KEY, trigramStringType());
-		return props;
-	}
+	
 
 }
