+++ elasticsearch/src/main/java/com/gentics/mesh/search/index/tagfamily/TagFamilyIndexHandler.java
@@ -29,6 +29,9 @@ public class TagFamilyIndexHandler extends AbstractIndexHandler<TagFamily> {
 	TagFamilyTransformer transformer;
 
 	@Inject
+	TagFamilyMappingProvider mappingProvider;
+
+	@Inject
 	public TagFamilyIndexHandler(SearchProvider searchProvider, Database db, BootstrapInitializer boot, SearchQueue searchQueue) {
 		super(searchProvider, db, boot, searchQueue);
 	}
@@ -38,8 +41,14 @@ public class TagFamilyIndexHandler extends AbstractIndexHandler<TagFamily> {
 		return TagFamily.class;
 	}
 
+	@Override
 	public TagFamilyTransformer getTransformer() {
 		return transformer;
+	}
+
+	@Override
+	public TagFamilyMappingProvider getMappingProvider() {
+		return mappingProvider;
 	}
 
 	@Override
@@ -49,8 +58,7 @@ public class TagFamilyIndexHandler extends AbstractIndexHandler<TagFamily> {
 
 	@Override
 	protected String composeIndexNameFromEntry(UpdateDocumentEntry entry) {
-		return TagFamily.composeIndexName(entry.getContext()
-				.getProjectUuid());
+		return TagFamily.composeIndexName(entry.getContext().getProjectUuid());
 	}
 
 	@Override
@@ -60,17 +68,14 @@ public class TagFamilyIndexHandler extends AbstractIndexHandler<TagFamily> {
 
 	@Override
 	public Completable store(TagFamily tagFamily, UpdateDocumentEntry entry) {
-		entry.getContext()
-				.setProjectUuid(tagFamily.getProject()
-						.getUuid());
+		entry.getContext().setProjectUuid(tagFamily.getProject().getUuid());
 		return super.store(tagFamily, entry);
 	}
 
 	@Override
 	public Map<String, String> getIndices() {
 		return db.tx(() -> {
-			ProjectRoot root = boot.meshRoot()
-					.getProjectRoot();
+			ProjectRoot root = boot.meshRoot().getProjectRoot();
 			Map<String, String> indexInfo = new HashMap<>();
 			for (Project project : root.findAllIt()) {
 				indexInfo.put(TagFamily.composeIndexName(project.getUuid()), TagFamily.TYPE);
@@ -93,8 +98,7 @@ public class TagFamilyIndexHandler extends AbstractIndexHandler<TagFamily> {
 
 	@Override
 	public RootVertex<TagFamily> getRootVertex() {
-		return boot.meshRoot()
-				.getTagFamilyRoot();
+		return boot.meshRoot().getTagFamilyRoot();
 	}
 
 }
