--- elasticsearch/src/main/resources/default-es-index-settings.json
@@ -1,17 +1,0 @@
-{
-  "analysis": {
-    "analyzer": {
-      "trigrams": {
-        "tokenizer": "my_ngram_tokenizer",
-        "filter": [ "lowercase" ]
-      }
-    },
-    "tokenizer": {
-      "my_ngram_tokenizer": {
-        "type": "nGram",
-        "min_gram": "3",
-        "max_gram": "3"
-      }
-    }
-  }
-}
